Create setupDevice.py
Running this program in terminal starts a server that you can access by typing http://127.0.0.1:8000 into a web browser. This will display 12 inputs and a submit button. You can type a word into each input and then press submit (or enter) and it will create a text file with each of those words (one input per line) and an mp3 file for each input. Those files will all be saved in the same location as this file, so you will probably want to locate this file in a folder. Either way, this file needs to be located at the same level as a folder called templates which needs to contain a file called form.html.