Create boot.py
This file is for the feather. If you connect a wire on your feather from D5 to ground, then you can write the mp3 files to the feather. In order to edit a file on the feather while it's connected to your computer, you need to disconnect the wire from D5. You have to plug/unplug the feather each time you connect/disconnect a wire to make this work.